Implemented pagination

+++ src/main/java/de/lukas/imagenetwork/controller/UserController.java
@@ -1,19 +1,18 @@
 package de.lukas.imagenetwork.controller;
 
 import de.lukas.imagenetwork.entity.User;
-import de.lukas.imagenetwork.exception.UserNotFoundException;
+import de.lukas.imagenetwork.enums.Role;
+import de.lukas.imagenetwork.exception.IncorrectUserException;
 import de.lukas.imagenetwork.model.UserCreate;
-import de.lukas.imagenetwork.repository.UserRepository;
 import de.lukas.imagenetwork.service.UserService;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-import java.time.Instant;
 import java.util.List;
-import java.util.Random;
-import java.util.logging.Logger;
+import java.util.Objects;
 
 @RestController
 @RequiredArgsConstructor
@@ -21,13 +20,19 @@ public class UserController {
 
     private final UserService userService;
 
-//    UserController(UserService userService) {
-//        this.userService = userService;
+//    @GetMapping("/users")
+//    List<User> all() {
+//        return userService.getAll();
 //    }
 
     @GetMapping("/users")
-    List<User> all() {
-        return userService.getAll();
+    Page<User> all(Pageable pageable) {
+        return userService.getAll(pageable);
+    }
+
+    @GetMapping("/user")
+    User currentUser(@AuthenticationPrincipal User user) {
+        return user;
     }
 
     @GetMapping("/user/{id}")
@@ -35,15 +40,15 @@ public class UserController {
         return userService.getById(id);
     }
 
-
     @PostMapping("/user")
     Long newUser(@RequestBody UserCreate userCreate) {
         return  userService.createUser(userCreate);
-
     }
 
     @DeleteMapping("/user/{id}")
-    String deleteUser(@PathVariable Long id) {
+    String deleteUser(@PathVariable Long id, @AuthenticationPrincipal User user) {
+        if(user.getRole() != Role.ADMIN && !Objects.equals(id, user.getId()))
+            throw new IncorrectUserException(user, id);
         userService.deleteUser(id);
         return "Done";
     }
